F08

+++ src/app/api/delete-chat/[id]/route.ts
@@ -3,13 +3,18 @@ import { eq } from "drizzle-orm";
 import { db } from "@/app/lib/db";
 import { chats, messages } from "@/app/lib/db/schema";
 import { deleteNamespace } from "@/app/lib/pinecone";
+import { deleteChatS3Files } from "@/app/lib/s3";
+import { auth } from "@clerk/nextjs/server";
 
 export const dynamic = "force-dynamic"; // <-- Indique qu'on force la route en mode dynamique
 
-export async function DELETE(
-  req: NextRequest,
-  { params }: { params: { id: string } }
-) {
+export async function DELETE(req: NextRequest,{ params }: { params: { id: string } }) {
+  const { userId } = await auth();
+
+  if (!userId) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  
   const { id } = await params;
   const chatId = parseInt(id, 10);
 
@@ -18,15 +23,19 @@ export async function DELETE(
   }
 
   try {
-    await db.delete(messages).where(eq(messages.chatId, chatId));
-
     const chatToDelete = await db.select().from(chats).where(eq(chats.id, chatId));
-    const fileKey = chatToDelete[0].fileKey
+  
+    const { fileKey, thumbnailUrl, userId: chatUserId } = chatToDelete[0];
+    if (chatUserId !== userId) {
+      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
+    }
 
-    const deletedChat = await db.delete(chats).where(eq(chats.id, chatId));
-    deleteNamespace(fileKey)
+    await db.delete(messages).where(eq(messages.chatId, chatId));
+    await await db.delete(chats).where(eq(chats.id, chatId));
+    await deleteNamespace(fileKey)
+    await deleteChatS3Files(fileKey, thumbnailUrl);
 
-    return NextResponse.json({ message: "[SUCCES] Chat supprimé", deletedChat });
+    return NextResponse.json({ message: "[SUCCES] Chat supprimé"});
   } catch (error) {
     return NextResponse.json(
       { error: "Internal server error", details: error },
